chore: make new webui header the default one [DET-5683]
